pyproject.toml
[project]
name = "blog-writer"
version = "0.1.0"
description = ""
authors = [
    {name = "juanarancibia",email = "arancibia.juan22@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "langchain[openai]>=0.3.20,<0.4.0",
    "langgraph>=0.3.11,<0.4.0",
    "langchain-community>=0.3.20,<0.4.0",
    "langchain-ollama>=0.2.3,<0.3.0",
    "langgraph-cli[inmem]>=0.1.75,<0.2.0",
    "tavily-python>=0.5.1,<0.6.0",
    "streamlit>=1.43.2,<2.0.0",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"